move config impl to cpp
