+++ cpp/src/io/utilities/config_utils.hpp
@@ -15,7 +15,6 @@
  */
 #pragma once
 
-#include <cstdlib>
 #include <string>
 
 namespace cudf::io::detail {
@@ -24,79 +23,33 @@ namespace cudf::io::detail {
  * @brief Returns the value of the environment variable, or a default value if the variable is not
  * present.
  */
-inline std::string getenv_or(std::string const& env_var_name, std::string_view default_val)
-{
-  auto const env_val = std::getenv(env_var_name.c_str());
-  return std::string{(env_val == nullptr) ? default_val : env_val};
-}
+std::string getenv_or(std::string const& env_var_name, std::string_view default_val);
 
 namespace cufile_integration {
-
-namespace {
-/**
- * @brief Defines which cuFile usage to enable.
- */
-enum class usage_policy : uint8_t { OFF, GDS, ALWAYS };
-
-/**
- * @brief Get the current usage policy.
- */
-inline usage_policy get_env_policy()
-{
-  static auto const env_val = getenv_or("LIBCUDF_CUFILE_POLICY", "GDS");
-  if (env_val == "OFF") return usage_policy::OFF;
-  if (env_val == "ALWAYS") return usage_policy::ALWAYS;
-  return usage_policy::GDS;
-}
-}  // namespace
 
 /**
  * @brief Returns true if cuFile and its compatibility mode are enabled.
  */
-inline bool is_always_enabled() { return get_env_policy() == usage_policy::ALWAYS; }
+bool is_always_enabled();
 
 /**
  * @brief Returns true if only direct IO through cuFile are enabled (compatiblity mode is disabled).
  */
-inline bool is_gds_enabled()
-{
-  return is_always_enabled() or get_env_policy() == usage_policy::GDS;
-}
+bool is_gds_enabled();
 
 }  // namespace cufile_integration
 
 namespace nvcomp_integration {
 
-namespace {
-/**
- * @brief Defines which nvCOMP usage to enable.
- */
-enum class usage_policy : uint8_t { OFF, STABLE, ALWAYS };
-
-/**
- * @brief Get the current usage policy.
- */
-inline usage_policy get_env_policy()
-{
-  static auto const env_val = getenv_or("LIBCUDF_NVCOMP_POLICY", "STABLE");
-  if (env_val == "OFF") return usage_policy::OFF;
-  if (env_val == "ALWAYS") return usage_policy::ALWAYS;
-  return usage_policy::STABLE;
-}
-}  // namespace
-
 /**
  * @brief Returns true if all nvCOMP uses are enabled.
  */
-inline bool is_all_enabled() { return get_env_policy() == usage_policy::ALWAYS; }
+bool is_all_enabled();
 
 /**
  * @brief Returns true if stable nvCOMP use is enabled.
  */
-inline bool is_stable_enabled()
-{
-  return is_all_enabled() or get_env_policy() == usage_policy::STABLE;
-}
+bool is_stable_enabled();
 
 }  // namespace nvcomp_integration
